# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Pessoas",
        "description": "Teste para a vaga de PHP Pleno",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pessoas": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Listar todos registros",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Cadastrar uma pessoa",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    },
                                    "nascionalidade": {
                                        "type": "string"
                                    },
                                    "sexo": {
                                        "enum": [
                                            "Masculino",
                                            "Feminino",
                                            "Outros"
                                        ]
                                    },
                                    "logradouro": {
                                        "type": "string"
                                    },
                                    "complemento": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "integer"
                                    },
                                    "bairro": {
                                        "type": "string"
                                    },
                                    "uf": {
                                        "type": "string"
                                    },
                                    "cidade": {
                                        "required": [
                                            "uf"
                                        ],
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "cep": "95873-377",
                                    "logradouro": "Rua Sebastião Escobar, 17. Bc. 7 Ap. 50",
                                    "numero": "4557",
                                    "complemento": "Apto 89",
                                    "bairro": "Rua Perez",
                                    "pais": "Brasil",
                                    "uf": "AM",
                                    "cidade": "Carrara do Leste"
                                }
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    },
                                    "nascionalidade": {
                                        "type": "string"
                                    },
                                    "sexo": {
                                        "type": "string",
                                        "enum": [
                                            "Masculino",
                                            "Feminino",
                                            "Outros"
                                        ]
                                    },
                                    "logradouro": {
                                        "type": "string"
                                    },
                                    "complemento": {
                                        "type": "string"
                                    },
                                    "bairro": {
                                        "type": "string"
                                    },
                                    "uf": {
                                        "type": "string"
                                    },
                                    "cidade": {
                                        "required": [
                                            "uf"
                                        ],
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nome": "Dr. Christopher Ricardo Ortega",
                                    "data_nascimento": "1988-10-14",
                                    "nascionalidade": "Americano",
                                    "cpf": "68204823533",
                                    "sexo": "Outros",
                                    "cep": "95873-377",
                                    "logradouro": "Rua Sebastião Escobar, 17. Bc. 7 Ap. 50",
                                    "numero": "4557",
                                    "complemento": "Apto 89",
                                    "bairro": "Rua Perez",
                                    "pais": "Brasil",
                                    "uf": "AM",
                                    "cidade": "Carrara do Leste"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
...
    }
}